foodapp/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.43.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.2
Django==5.1.6
django-celery-beat==2.8.0
django-ckeditor==6.7.2
django-filter==25.1
django-js-asset==3.0.1
django-oauth-toolkit==3.0.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.15.2
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
kombu==5.5.3
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.51
pycparser==2.22
PyMySQL==1.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
